统一时间间隔参数为浮点数（秒），优化相关函数以提高代码一致性和可读性；更新主程序中的时间计算逻辑，确保精度提升。

+++ lib/servo/src/core/monitor.cc
@@ -19,9 +19,9 @@ void Monitor::UseMonitoring(Print &serial) { monitorPort_ = &serial; }
 
 void Monitor::LinkMotor(Servo *servo) { servo_ = servo; }
 
-void Monitor::Process(uint32_t dt) {
+void Monitor::Process(float dt) {
   if (!monitorPort_) return;
-  static uint32_t accumulated_time = 0;
+  static float accumulated_time = 0;
   accumulated_time += dt;
   if (accumulated_time < period_) return;
   accumulated_time = 0;
