+++ lib/servo/src/inst/inst_serial_transport.h
@@ -36,10 +36,10 @@ class InstSerialTransport : public InstTransportInterface {
 
   /**
    * @brief 处理数据
-   * @param dt 时间间隔(微秒)
+   * @param dt 时间间隔(秒)
    * @return 错误码
    */
-  Error Process(const uint32_t dt) override;
+  Error Process(const float dt) override;
 
   /**
    * @brief 发送数据
